drivers: sensor: icm42605: sign-extend FIFO temperature byte
The 8-bit FIFO temperature field is two's complement, but the driver
cast the uint8_t FIFO byte directly to int16_t, zero-extending it and
turning every reading below 25 degC into a bogus 87..148 degC value.
Cast through int8_t so the sign bit is preserved.

+++ drivers/sensor/tdk/icm42605/icm42605.c
@@ -206,7 +206,7 @@ static int icm42605_sample_fetch(const struct device *dev,
 			}
 			if (!(drv_data->fifo_data[0] & BIT_FIFO_HEAD_GYRO)) {
 				drv_data->temp =
-					(int16_t)(drv_data->fifo_data[7]);
+					(int16_t)(int8_t)drv_data->fifo_data[7];
 			} else {
 				if (!(drv_data->fifo_data[7] ==
 				      FIFO_GYRO0_RESET_VALUE &&
@@ -223,7 +223,7 @@ static int icm42605_sample_fetch(const struct device *dev,
 						+ (drv_data->fifo_data[12]);
 				}
 				drv_data->temp =
-					(int16_t)(drv_data->fifo_data[13]);
+					(int16_t)(int8_t)drv_data->fifo_data[13];
 			}
 		} else {
 			if (drv_data->fifo_data[0] & BIT_FIFO_HEAD_GYRO) {
@@ -242,7 +242,7 @@ static int icm42605_sample_fetch(const struct device *dev,
 						+ (drv_data->fifo_data[6]);
 				}
 				drv_data->temp =
-					(int16_t)(drv_data->fifo_data[7]);
+					(int16_t)(int8_t)drv_data->fifo_data[7];
 			}
 		}
 	}
